Move AddSensors container folder under containers
