+++ packages/webapp/src/components/Sensor/v2/Partners.tsx
@@ -21,13 +21,13 @@ import { validate as uuidValidate } from 'uuid';
 import Input, { getInputErrors } from '../../Form/Input';
 import InputBaseLabel from '../../Form/InputBase/InputBaseLabel';
 import { Main } from '../../Typography';
-import { PARTNERS } from '../../../containers/LocationDetails/PointDetails/SensorDetail/v2/constants';
+import { PARTNERS } from '../../../containers/AddSensors/constants';
 import { ReactComponent as ExternalLinkIcon } from '../../../assets/images/icon_external_link.svg';
 import {
   AddSensorsFormFields,
   FarmAddonField,
   PARTNER,
-} from '../../../containers/LocationDetails/PointDetails/SensorDetail/v2/types';
+} from '../../../containers/AddSensors/types';
 import styles from './styles.module.scss';
 import { AddonPartner } from '../../../types';
 import { createSensorsUrl } from '../../../util/siteMapConstants';
