+++ packages/webapp/src/containers/AddSensors/Partners.tsx
@@ -12,8 +12,8 @@
  *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  *  GNU General Public License for more details, see <https://www.gnu.org/licenses/>.
  */
-import PurePartners from '../../../../../components/Sensor/v2/Partners';
-import { useGetFarmAddonQuery } from '../../../../../store/api/apiSlice';
+import PurePartners from '../../components/Sensor/v2/Partners';
+import { useGetFarmAddonQuery } from '../../store/api/apiSlice';
 import { PARTNERS } from './constants';
 
 const Partners = ({
