+++ packages/webapp/src/containers/AddSensors/index.tsx
@@ -17,14 +17,14 @@ import { useTranslation } from 'react-i18next';
 import { useDispatch } from 'react-redux';
 import { FieldValues } from 'react-hook-form';
 import { History } from 'history';
-import { ContextForm, Variant } from '../../../../../components/Form/ContextForm';
+import { ContextForm, Variant } from '../../components/Form/ContextForm';
 import Partners from './Partners';
-import PageTitle from '../../../../../components/PageTitle/v2';
-import { enqueueErrorSnackbar } from '../../../../Snackbar/snackbarSlice';
-import { useAddFarmAddonMutation, useLazyGetSensorsQuery } from '../../../../../store/api/apiSlice';
+import PageTitle from '../../components/PageTitle/v2';
+import { enqueueErrorSnackbar } from '../Snackbar/snackbarSlice';
+import { useAddFarmAddonMutation, useLazyGetSensorsQuery } from '../../store/api/apiSlice';
 import { type AddSensorsFormFields, FarmAddonField, PARTNER } from './types';
 import { PARTNERS } from './constants';
-import { createSensorsUrl } from '../../../../../util/siteMapConstants';
+import { createSensorsUrl } from '../../util/siteMapConstants';
 import styles from './styles.module.scss';
 
 interface AddSensorProps {
